+++ NEWS.md
@@ -1,3 +1,16 @@
+# semFromKeys 0.3.0
+
+* Added the `sem.cor()` function, which takes CFA outputs and produces a correlation matrix between latent variables and, optionally, single items.
+
+# semFromKeys 0.2.4
+
+* Updated `cache.clean()` to only work in non-interactive sessions when `interactive = FALSE`.
+* Changed `cache.setup()` and `cache.clean()` examples to `\dontrun` to prevent saved elements from remaining after CRAN checks of `\donttest`.
+
+# semFromKeys 0.2.3
+
+* Fixed esem.from.mods returning an empty `fit_measures` entry when `fit_save = FALSE`.
+
 # semFromKeys 0.2.2
 
 * Fixed the option issue that remained at the last CRAN submission (in esem.from.mods.R).
@@ -15,7 +28,7 @@
 
 # semFromKeys 0.1.1
 
-* Added caching with [cache.setup()] and [cache.clean()] for use with
+* Added caching with `cache.setup()` and `cache.clean()` for use with
 `save_out = TRUE` and `check = TRUE`.
 
 # semFromKeys 0.1.0